Bug Float

+++ runs/mqttsub.py
@@ -1407,7 +1407,7 @@ def on_message(client, userdata, msg):
                     f.write(msg.payload.decode("utf-8"))
                     f.close()
             if (msg.topic == "openWB/set/houseBattery/WhImported"):
-                if (int(msg.payload) >= 0 and int(msg.payload) <= 9000000):
+                if (float(msg.payload) >= 0 and float(msg.payload) <= 9000000):
                     f = open('/var/www/html/openWB/ramdisk/speicherikwh', 'w')
                     f.write(msg.payload.decode("utf-8"))
                     f.close()
